+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Fixed bug where keyword arguments may not work for All Time Teams stat sheet in Excel
 - Fixed bug where leagues with some League Median years and some non League Median years would not save to excel
+- Improved Excel tests
 
 ## [1.11.0]